+++ .eslintrc.js
@@ -53,9 +53,10 @@ module.exports = {
 				devDependencies: [
 					'tools/**/*.js',
 					'**/*.test.js',
-					'gatsby-config.js',
-					'gatsby-ssr.js',
+					'**/gatsby-config.js',
+					'**/gatsby-ssr.js',
 					'testsSetup.js',
+					'jestPreprocess.js',
 					'*.config.js',
 				],
 			},