+++ demo/src/components/index.js
@@ -1,9 +1,4 @@
 export { default as ZoeLogo } from './ZoeLogo';
-export {
-	default as DevConsole,
-	useDevConsole,
-	DevConsoleProvider,
-} from './DevConsole';
 export { default as OptionalFormattedMessage } from './OptionalFormattedMessage';
 export { default as LoanFormSection } from './LoanFormSection';
 export { default as BorrowersFormSection } from './BorrowersFormSection';