+++ demo/src/utils.js
@@ -1,9 +1,3 @@
-import { filter, keys, o } from 'ramda';
-
-export const round = (x) => Math.round(x * 100) / 100;
-
-export const getTruthyKeys = o(keys, filter(Boolean));
-
 // https://stackoverflow.com/questions/179355/clearing-all-cookies-with-javascript
 export const deleteAllCookies = () => {
 	const cookies = document.cookie.split(';');